+++ megazords/ios-rust/tests/MozillaRustComponentsTests/SyncManagerTelemetryTests.swift
@@ -274,4 +274,20 @@ class SyncManagerTelemetryTests: XCTestCase {
                                   submitCreditCardsPing: submitCreditCardsPing,
                                   submitTabsPing: submitTabsPing)
     }
+
+    func testReceivesOpenSyncSettingsMenuTelemetry() {
+        SyncManagerComponent.reportOpenSyncSettingsMenuTelemetry()
+        let events = GleanMetrics.SyncSettings.openMenu.testGetValue()!
+        XCTAssertEqual(1, events.count)
+    }
+
+    func testReceivesSaveSyncSettingsTelemetry() {
+        let enabledEngines = ["bookmarks", "tabs"]
+        let disabledEngines = ["logins"]
+        SyncManagerComponent.reportSaveSyncSettingsTelemetry(enabledEngines: enabledEngines, disabledEngines: disabledEngines)
+        let events = GleanMetrics.SyncSettings.save.testGetValue()!
+        XCTAssertEqual(1, events.count)
+        XCTAssertEqual("bookmarks,tabs", events[0].extra!["enabled_engines"])
+        XCTAssertEqual("logins", events[0].extra!["disabled_engines"])
+    }
 }